points_service/requirements.txt
# points_service/requirements.txt

"""
Назначение: зависимости Python для сервиса.
Используется: в Dockerfile/pip install.
"""

fastapi
uvicorn
sqlalchemy
asyncpg
python-jose
pydantic-settings
